+++ 2D/Potts.cpp
@@ -8,7 +8,7 @@ using namespace std;
 namespace vb {
     template <> Color to_Color(int t) {
         static const vector<Color> C{RED, BLUE, YELLOW, GREEN, MAGENTA, CYAN};
-        if ((t >= 0) && (t < 6)) return C[t];
+        if ((t >= 0) && (t < 6)) return C[size_t(t)];
         if (t < 0) return WHITE;
         return Indexed(t);
     }
